Fixed problem with timeVal() function
had forgotten to initialize updateTime to zero at startup (how did this ever work?)

+++ src/HomeSpan.cpp
@@ -1556,7 +1556,7 @@ void SpanCharacteristic::setVal(double val){
 
 ///////////////////////////////
 
-int SpanCharacteristic::timeVal(){
+unsigned long SpanCharacteristic::timeVal(){
   
   return(homeSpan.snapTime-updateTime);
 }
